🎨 Fix stories
